requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
blinker==1.6.2
celery==5.3.4
class-registry==2.1.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.2
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Markdown==3.5
MarkupSafe==2.1.3
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
regex==2023.10.3
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.9
Werkzeug==2.3.6
